+++ NEWS.md
@@ -1,8 +1,20 @@
+## 0.4.5 
+
+* Apply label information to `epiDisplay.object`: `LabelepiDisplay` function
+
+* Apply label information to `lmer.display`, `coxme.display`: `LabeljsMixed` function
+
+
+
+
 ## 0.4.0
 
 * Table from `coxme.object` (**coxme** package)
 
 * Error correct: 1 variable case.
+
+
+
 
 ## 0.3.5